Update features and deployment configuration
