+++ backend/.env
@@ -1,3 +1,4 @@
 PORT=5000
 MONGO_URI=mongodb+srv://abhaynarayan0001_db_user:Cb2xDn7fg9XX9LFW@selfpiebackend.s4bdbnp.mongodb.net/?appName=selfpieBackend
 GEMINI_API_KEY=AIzaSyAVp1DGi9Pq3xsAmdQvcrjvc7Glt-M5XmM
+JWT_SECRET=BG3AVwhsJCRnfhkQwD8dvPlUhbDxiqoY
