+++ backend/.env.example
@@ -1,0 +1,4 @@
+PORT=5000
+MONGO_URI=mongodb_uri_here
+JWT_SECRET=your_jwt_secret_here
+GEMINI_API_KEY=your_gemini_api_key_here
