+++ frontend/src/store/SpAbhay_useCartStore.js
@@ -4,6 +4,16 @@ export const useCartStore = create((set) => ({
   cartItems: [],
   merchantOrders: [], 
   currentShopId: null,
+  activeOrderId: localStorage.getItem('activeOrderId') || null,
+
+  setActiveOrderId: (id) => {
+    if (id) {
+       localStorage.setItem('activeOrderId', id);
+    } else {
+       localStorage.removeItem('activeOrderId');
+    }
+    set({ activeOrderId: id });
+  },
 
   setShopId: (id) => set({ currentShopId: id }),
